Move inmemory channel to pkg/provisioners
