+++ pkg/provisioners/inmemory/controller/main.go
@@ -18,10 +18,10 @@ package main
 
 import (
 	"flag"
+	"github.com/knative/eventing/pkg/provisioners/inmemory/channel"
 
 	eventingv1alpha1 "github.com/knative/eventing/pkg/apis/eventing/v1alpha1"
-	"github.com/knative/eventing/pkg/reconciler/v1alpha1/inmemory/channel"
-	"github.com/knative/eventing/pkg/reconciler/v1alpha1/inmemory/clusterchannelprovisioner"
+	"github.com/knative/eventing/pkg/provisioners/inmemory/clusterchannelprovisioner"
 	"github.com/knative/eventing/pkg/provisioners"
 	istiov1alpha3 "github.com/knative/pkg/apis/istio/v1alpha3"
 	"github.com/knative/pkg/signals"
